+++ ubo_app/utils/__init__.py
@@ -1,5 +1,4 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107
-import subprocess
 from pathlib import Path
 
 from headless_kivy_pi.constants import BYTES_PER_PIXEL
@@ -35,7 +34,6 @@ def turn_off_screen() -> None:
     display = HeadlessWidget._display  # noqa: SLF001
     data = [0] * HeadlessWidget.width * HeadlessWidget.height * BYTES_PER_PIXEL
     display._block(0, 0, HeadlessWidget.width - 1, HeadlessWidget.height - 1, data)  # noqa: SLF001
-    subprocess.run(['/usr/bin/env', 'systemctl', 'poweroff', '-i'], check=True)  # noqa: S603
 
 
 def turn_on_screen() -> None: